update add router
