+++ .gitignore
@@ -25,3 +25,6 @@ dist
 tmp
 
 .firebase
+
+# routify
+.routify
