+++ package.json
@@ -17,5 +17,8 @@
     "tslib": "^2.6.2",
     "typescript": "^5.2.2",
     "vite": "^5.2.0"
+  },
+  "dependencies": {
+    "@roxi/routify": "^3.0.0-next.227"
   }
 }
